Add security enhancements around mediated deposit
Mediated deposit is inherently insecure, and involves a trust relationship between the depositor and the mediated user.  This is fine for basic deposit, but when we introduce CRUD there are more serious security challenges.  In order to temporarily manage these pending a more
substantive solution, updates to items via mediated interactions can be restricted to a set of permitted mediators, set in configuration.  This means, anyone can create a deposit on behalf of another user, but only one of the specified mediators can make any changes to it after that.
This is back-compat with the approach used in SWORDv1 and introduces a degree of additional security while a better overall solution is found.